small change: tpye

+++ firmware/IR-station/IR_op.h
@@ -8,12 +8,11 @@
 #define IR_STATION_MODE_STA   1
 #define IR_STATION_MODE_AP    2
 
-// define variables
 extern remocon ir[IR_CH_SIZE];
 extern uint8_t mode;
 
 void modeSetup(void);
-void setMode(int newMode);
+void setMode(uint8_t newMode);
 
 void irSendSignal(int ch);
 int irRecodeSignal(int ch);
